requirements.txt
Flask
gunicorn
tensorflow-cpu
sklearn
numpy
pandas